Merge in WLMDB/wlmdb from DBS-2801_multiinstance to master

* commit 'f51d38bb80ee86ba6cd47f15d0044159c7f80f14':
  DBS-2801: Review comments
  DBS-2801: Refactoring
  DBS-2801: resource-credentials changes

+++ server/src/operations/discover-operations.ts
@@ -3,7 +3,7 @@ import config from 'config';
 
 import { STORAGE_TYPE, JOBSTATUS, JOBTYPE } from '@prisma/client';
 import { FileSystem, FileSystemType } from '@aws-sdk/client-fsx';
-import { attempt, compact, uniqBy, isEmpty } from 'lodash-es';
+import { attempt, compact, uniqBy, isEmpty, cloneDeep } from 'lodash-es';
 import { DescribeInstancesCommandInput, InstanceStateName, Vpc } from '@aws-sdk/client-ec2';
 import { CommandInvocationStatus, ConnectionStatus } from '@aws-sdk/client-ssm';
 import throat from 'throat';
@@ -611,8 +611,8 @@ function prepareParametersToStore(instanceId: string, credentials: DiscoverCrede
         if (resourceType === RESOURCESTYPE.MSSQL) {
             const sqlItem = acc.find(el => el.value.sql);
 
-            if (sqlItem && Array.isArray(sqlItem)) {
-                sqlItem.push({
+            if (sqlItem && Array.isArray(sqlItem.value.sql)) {
+                sqlItem.value.sql.push({
                     sqlinstancename: resourceId,
                     username,
                     password
@@ -1198,8 +1198,6 @@ async function validateCredentials(
 ) {
     logger.info('validateCredentials', { instanceId, fsxCredentials, sqlCredentials });
 
-    await verifyAndCreateCredentials(credentialsId, region, instanceId, fsxCredentials, sqlCredentials);
-
     const connectionStatus = await getSSMConnectionStatus(credentialsId, region, instanceId);
     if (connectionStatus.Status !== ConnectionStatus.CONNECTED) {
         const errorMessage = `Unable to validate the credentials through SSM, for host ${instanceId}`;
@@ -1211,6 +1209,10 @@ async function validateCredentials(
         ]);
         throw createError(HttpErrorCodes.INTERNAL_SERVER_ERROR, errorMessage);
     }
+
+    const newSqlCredentials = cloneDeep(sqlCredentials);
+    await verifyAndCreateCredentials(credentialsId, region, instanceId, fsxCredentials, newSqlCredentials);
+
     let parsedResponse;
 
     try {
@@ -1253,6 +1255,7 @@ async function validateCredentials(
 
         const response: Record<string, string> = {};
         const paramesToDelete: string[] = [];
+        const instancesToBeDeleted: string[] = [];
 
         if (parsedResponse.requiredModuleError) {
             response.requiredModuleError = parsedResponse.requiredModuleError;
@@ -1264,7 +1267,7 @@ async function validateCredentials(
 
             if (sqlCredentials.length) {
                 if (parsedResponse.sqlInstanceConnectivity === false) {
-                    paramesToDelete.push(`${SSM_PARAM_PREFIX}${instanceId}`);
+                    instancesToBeDeleted.push(sqlCredentials[0].resourceId);
                     response.sqlServerError = parsedResponse?.sqlerror;
                 } else if (parsedResponse.sqlInstanceConnectivity === true) {
                     response.sqlServerEdition = parsedResponse?.sqlEdition;
@@ -1273,8 +1276,23 @@ async function validateCredentials(
             }
         }
 
-        await deleteSSMParameter(credentialsId, region, paramesToDelete);
-
+        if (instancesToBeDeleted.length > 0) {
+            if (newSqlCredentials.length === instancesToBeDeleted.length) {
+                // Delete the parameter store all credentials are invalid
+                paramesToDelete.push(`${SSM_PARAM_PREFIX}${instanceId}`);
+                await deleteSSMParameter(credentialsId, region, paramesToDelete);
+            } else {
+                // Rewrite parameter store after removing invalid credentials
+                const latestSqlCredentials = newSqlCredentials.filter(
+                    e => !instancesToBeDeleted.includes(e.resourceId)
+                );
+                const creds = prepareParametersToStore(instanceId, [
+                    ...(fsxCredentials ? [fsxCredentials] : []),
+                    ...latestSqlCredentials
+                ]);
+                await ssmPutParameters(credentialsId, region, creds);
+            }
+        }
         return response;
     } catch (error: any) {
         // delete the ssm parameters if its already created
@@ -1288,7 +1306,7 @@ async function validateCredentials(
         await deleteSSMParameter(credentialsId, region, paramesToDelete);
         throw createError(
             HttpErrorCodes.INTERNAL_SERVER_ERROR,
-            `unable to validate the credentials . Reason: ${error?.message}.`
+            `Unable to validate the credentials . Reason: ${error?.message}.`
         );
     }
 }
@@ -1315,10 +1333,22 @@ async function verifyAndCreateCredentials(
     }
 
     if (sqlCredentials.length) {
-        const SSMParameters = await getParameter(credentialsId, region, `${SSM_PARAM_PREFIX}${instanceId}`);
-        if (!SSMParameters) {
+        const existingParameters = await getParameter(credentialsId, region, `${SSM_PARAM_PREFIX}${instanceId}`);
+        if (!existingParameters) {
             const newSSMParameters: string[] = await getAsyncLocalStorageResource(NEW_SSM_PARAMETERS);
             setAsyncLocalStorageResource(NEW_SSM_PARAMETERS, [...(newSSMParameters || []), instanceId]);
+        } else {
+            const { sql } = JSON.parse(existingParameters);
+            if (sql) {
+                sql.forEach((e: { sqlinstancename: string; username: string; password: string }) => {
+                    sqlCredentials.push({
+                        resourceId: e.sqlinstancename,
+                        resourceType: RESOURCESTYPE.MSSQL,
+                        username: e.username,
+                        password: e.password
+                    });
+                });
+            }
         }
     }
